Should work properly now

+++ src/TheHolyCompiler-core/core/compiler/compiler.h
@@ -450,7 +450,7 @@ private:
 	utils::List<utils::String> defines;
 	utils::List<utils::String> includes;
 
-	utils::List<uint32> locations;
+	utils::List<uint64> locations;
 
 	utils::List<parsing::Token> Tokenize();
 	void ParseTokens(utils::List<parsing::Token>& tokens);
